+++ example/integration_test/create_pdf_from_multiple_images_integration_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
+import 'package:pdf_combiner/exception/pdf_combiner_exception.dart';
 import 'package:pdf_combiner/pdf_combiner.dart';
-import 'package:pdf_combiner/responses/pdf_combiner_status.dart';
 
 import 'test_file_helper.dart';
 
@@ -24,39 +24,47 @@ void main() {
         outputPath: outputPath,
       );
 
-      expect(result.status, PdfCombinerStatus.success);
       expect(result.outputPath, '${TestFileHelper.basePath}/merged_output.pdf');
       expect(result.message, 'Processed successfully');
     }, timeout: Timeout.none);
 
     testWidgets('Test creating pdf with empty list', (tester) async {
-      final result = await PdfCombiner.createPDFFromMultipleImages(
-        inputPaths: [],
-        outputPath: '${TestFileHelper.basePath}/assets/merged_output.pdf',
+      expect(
+        () => PdfCombiner.createPDFFromMultipleImages(
+          inputPaths: [],
+          outputPath: '${TestFileHelper.basePath}/assets/merged_output.pdf',
+        ),
+        throwsA(
+          predicate(
+            (e) =>
+                e is PdfCombinerException &&
+                e.message == 'The parameter (inputPaths) cannot be empty',
+          ),
+        ),
       );
-
-      expect(result.status, PdfCombinerStatus.error);
-      expect(result.outputPath, "");
-      expect(result.message, 'The parameter (inputPaths) cannot be empty');
     }, timeout: Timeout.none);
 
     testWidgets('Test creating pdf with non-existing file', (tester) async {
       final helper = TestFileHelper([]);
       final inputPaths = await helper.prepareInputFiles();
-
-      // Add a non-existing file path manually
       inputPaths.add('${TestFileHelper.basePath}/assets/non_existing.jpg');
       final outputPath = await helper.getOutputFilePath('merged_output.pdf');
 
-      final result = await PdfCombiner.createPDFFromMultipleImages(
-        inputPaths: inputPaths,
-        outputPath: outputPath,
+      expect(
+        () => PdfCombiner.createPDFFromMultipleImages(
+          inputPaths: inputPaths,
+          outputPath: outputPath,
+        ),
+        throwsA(
+          predicate(
+            (e) =>
+                e is PdfCombinerException &&
+                e.message.startsWith(
+                  'File is not an image or does not exist:',
+                ),
+          ),
+        ),
       );
-
-      expect(result.status, PdfCombinerStatus.error);
-      expect(result.outputPath, "");
-      expect(result.message,
-          startsWith('File is not an image or does not exist:'));
     }, timeout: Timeout.none);
 
     testWidgets('Test creating pdf with non-supported file', (tester) async {
@@ -65,15 +73,20 @@ void main() {
       final inputPaths = await helper.prepareInputFiles();
       final outputPath = await helper.getOutputFilePath('merged_output.pdf');
 
-      final result = await PdfCombiner.createPDFFromMultipleImages(
-        inputPaths: inputPaths,
-        outputPath: outputPath,
+      expect(
+        () => PdfCombiner.createPDFFromMultipleImages(
+          inputPaths: inputPaths,
+          outputPath: outputPath,
+        ),
+        throwsA(
+          predicate(
+            (e) =>
+                e is PdfCombinerException &&
+                e.message ==
+                    'File is not an image or does not exist: ${TestFileHelper.basePath}/document_1.pdf',
+          ),
+        ),
       );
-
-      expect(result.status, PdfCombinerStatus.error);
-      expect(result.outputPath, "");
-      expect(result.message,
-          'File is not an image or does not exist: ${TestFileHelper.basePath}/document_1.pdf');
     }, timeout: Timeout.none);
   });
 }