+++ lib/responses/image_from_pdf_response.dart
@@ -1,10 +1,5 @@
-import 'package:pdf_combiner/responses/pdf_combiner_status.dart';
-
 /// Represents the response for extracting images from a PDF file.
 class ImageFromPDFResponse {
-  /// The status of the image extraction process.
-  PdfCombinerStatus status;
-
   /// An optional message providing additional details about the process.
   String? message;
 
@@ -13,16 +8,14 @@ class ImageFromPDFResponse {
 
   /// Creates a response object for image extraction from a PDF.
   ///
-  /// - [status] The status of the extraction process (required).
   /// - [outputPaths] A list of file paths for the extracted images (defaults to an empty list).
   /// - [message] An optional message with additional information.
   ImageFromPDFResponse({
-    required this.status,
     this.outputPaths = const [],
     this.message,
   });
 
   @override
   String toString() =>
-      "ImageFromPDFResponse{outputPaths: $outputPaths, message: $message, status: $status }";
+      "ImageFromPDFResponse{outputPaths: $outputPaths, message: $message}";
 }