+++ test/pdf_combiner_generate_from_documents_test.dart
@@ -6,7 +6,6 @@ import 'package:pdf_combiner/communication/pdf_combiner_method_channel.dart';
 import 'package:pdf_combiner/communication/pdf_combiner_platform_interface.dart';
 import 'package:pdf_combiner/exception/pdf_combiner_exception.dart';
 import 'package:pdf_combiner/pdf_combiner.dart';
-import 'package:pdf_combiner/responses/pdf_combiner_status.dart';
 import 'package:pdf_combiner/utils/document_utils.dart';
 
 import 'mocks/mock_pdf_combiner_platform.dart';
@@ -60,11 +59,10 @@ void main() {
       );
 
       // Verify the result matches the expected mock values.
-      expect(result.status, PdfCombinerStatus.success);
       expect(result.outputPath, "output/path.pdf");
       expect(result.message, 'Processed successfully');
       expect(result.toString(),
-          'GeneratePdfFromDocumentsResponse{outputPath: ${result.outputPath}, message: ${result.message}, status: ${result.status} }');
+          'GeneratePdfFromDocumentsResponse{outputPath: ${result.outputPath}, message: ${result.message}}');
     });
 
     // Test for successfully combining multiple PDFs using PdfCombiner.
@@ -86,11 +84,10 @@ void main() {
       );
 
       // Verify the result matches the expected mock values.
-      expect(result.status, PdfCombinerStatus.success);
       expect(result.outputPath, "path.pdf");
       expect(result.message, 'Processed successfully');
       expect(result.toString(),
-          'GeneratePdfFromDocumentsResponse{outputPath: ${result.outputPath}, message: ${result.message}, status: ${result.status} }');
+          'GeneratePdfFromDocumentsResponse{outputPath: ${result.outputPath}, message: ${result.message}}');
     });
 
     // Test for error with wrong outputPath in combining multiple PDFs using PdfCombiner.